[MISC] Rename core/detail/to_string.hpp to core/debug_stream/detail/to_string.hpp
