+++ include/seqan3/core/detail/to_string.hpp
@@ -7,31 +7,14 @@
 
 /*!\file
  * \author Svenja Mehringer <svenja.mehringer AT fu-berlin.de>
- * \brief Auxiliary for pretty printing of exception messages.
+ * \brief [DEPRECATED] Auxiliary for pretty printing of exception messages.
+ * \deprecated This header is deprecated and will be removed in SeqAn-3.1.0;. Please \#include
+ *             <seqan3/core/debug_stream/detail/to_string.hpp> instead.
  */
 
 #pragma once
 
-#include <sstream>
+#include <seqan3/core/debug_stream/detail/to_string.hpp>
 
-#include <seqan3/core/detail/debug_stream_type.hpp>
-
-namespace seqan3::detail
-{
-
-/*!\brief Streams all parameters via the seqan3::debug_stream and returns a concatenated string.
- *
- * \tparam    value_type Must be streamable (stream << value).
- * \param[in] values     Variable number of parameters of any type that implement the stream operator.
- * \returns A concatenated string of all values (no separator in between is added).
- */
-template <typename ...value_type>
-std::string to_string(value_type && ...values)
-{
-    std::stringstream stream;
-    debug_stream_type dstream{stream};
-    (dstream << ... << std::forward<value_type>(values));
-    return stream.str();
-}
-
-} // namespace seqan3::detail
+SEQAN3_DEPRECATED_HEADER(
+   "This header is deprecated and will be removed in SeqAn-3.1.0; Please #include <seqan3/core/debug_stream/detail/to_string.hpp> instead.")
