+++ include/seqan3/io/alignment_file/format_bam.hpp
@@ -12,7 +12,9 @@
 
 #pragma once
 
+#include <seqan3/std/concepts>
 #include <iterator>
+#include <seqan3/std/ranges>
 #include <string>
 #include <vector>
 
@@ -22,8 +24,8 @@
 #include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/concept/tuple.hpp>
+#include <seqan3/core/debug_stream/detail/to_string.hpp>
 #include <seqan3/core/detail/debug_stream_optional.hpp>
-#include <seqan3/core/detail/to_string.hpp>
 #include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/io/alignment_file/detail.hpp>
@@ -42,8 +44,6 @@
 #include <seqan3/range/views/slice.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
 #include <seqan3/range/views/take_until.hpp>
-#include <seqan3/std/concepts>
-#include <seqan3/std/ranges>
 
 namespace seqan3
 {
