+++ include/seqan3/io/alignment_file/format_sam.hpp
@@ -12,14 +12,17 @@
 
 #pragma once
 
+#include <seqan3/std/algorithm>
+#include <seqan3/std/concepts>
 #include <iterator>
+#include <seqan3/std/ranges>
 #include <string>
 #include <vector>
 
 #include <seqan3/core/char_operations/predicate.hpp>
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/concept/tuple.hpp>
-#include <seqan3/core/detail/to_string.hpp>
+#include <seqan3/core/debug_stream/detail/to_string.hpp>
 #include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/io/alignment_file/detail.hpp>
@@ -43,9 +46,6 @@
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/range/views/to_char.hpp>
 #include <seqan3/range/views/to.hpp>
-#include <seqan3/std/algorithm>
-#include <seqan3/std/concepts>
-#include <seqan3/std/ranges>
 
 namespace seqan3
 {
