+++ test/include/seqan3/test/expect_range_eq.hpp
@@ -17,8 +17,8 @@
 #include <seqan3/std/algorithm>
 #include <seqan3/std/ranges>
 
+#include <seqan3/core/debug_stream/detail/to_string.hpp>
 #include <seqan3/core/detail/debug_stream_range.hpp>
-#include <seqan3/core/detail/to_string.hpp>
 #include <seqan3/test/pretty_printing.hpp>
 
 namespace seqan3::test
